ref: Rename undefined deployment as null
