+++ src/deployments/index.js
@@ -1,1 +1,1 @@
-export * from "./deployment";
+export * from "./deployments";
