--- src/deployments/undefined/index.js
@@ -1,1 +1,0 @@
-export * from "./undefined";
